prepared infra

+++ src/main/kotlin/com/watchdognet/catalog_service/application/product/update/UpdateProductCommand.kt
@@ -1,10 +1,12 @@
 package com.watchdognet.catalog_service.application.product.update
 
+import java.math.BigDecimal
+
 data class UpdateProductCommand(
   val id: String,
   val name: String,
   val description: String,
-  val price: Double,
+  val price: BigDecimal,
   val stock: Int,
   val active: Boolean
 )
